Commit after resolving conflict on doing rebase

+++ src/main/java/com/microservices/posts/Service/PostsService.java
@@ -1,74 +1,52 @@
 package com.microservices.posts.Service;
 
-import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.model.*;
+import com.amazonaws.services.sns.AmazonSNS;
+import com.amazonaws.services.sns.model.PublishRequest;
+import com.amazonaws.services.sns.model.PublishResult;
+import com.google.gson.Gson;
 import com.microservices.posts.Entity.Posts;
 import com.microservices.posts.Repository.PostsRepository;
-import com.microservices.posts.dto.DownloadFileDto;
 import com.microservices.posts.dto.UpdatePostDto;
+import com.microservices.posts.dto.awsSns.AwsSnsPostDto;
 import com.microservices.posts.models.CreatePostRequest;
 import com.microservices.posts.models.ResponsePosts;
-import com.microservices.posts.utils.FileUtils;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
 
-import java.io.File;
-import java.io.IOException;
 import java.util.*;
 
 @Service
 public class PostsService {
     private final PostsRepository postsRepository;
-    private final AmazonS3 amazonS3Client;
+    final private AmazonSNS amazonSNSClient;
+    @Value("${aws.snsArn}")
+    private String snsArn;
+    @Value("${aws.s3Bucket}")
+    private String AWS_BUCKET_NAME;
+
     public PostsService(PostsRepository postsRepository,
-                        AmazonS3 amazonS3Client){
+                        AmazonSNS amazonSNSClient){
         this.postsRepository = postsRepository;
-        this.amazonS3Client = amazonS3Client;
+        this.amazonSNSClient = amazonSNSClient;
     }
-    private final String AWS_BUCKET_NAME = "psotmicroservicebucket";
+
     public String createPost(CreatePostRequest request, String userName){
         Posts posts = new Posts(request.getTitle(), request.getContent(), userName);
         Posts createdPost = postsRepository.save(posts);
         System.out.println("created post id : " + createdPost.getId());
-        return createdPost.getId();
-    }
 
-    public void uploadFile(String postId, MultipartFile file){
-        File convertedFile = FileUtils.convertMultipartToFile(file);
-        String fileName = FileUtils.generateFileName(file);
-        String filePath = "spring_microservice/" +fileName;
-        UpdatePostDto updatePostDto = new UpdatePostDto(filePath);
-
-        Posts post = getPost(postId);
-
-        if (post.getImageId() != null && !post.getImageId().trim().isEmpty()){
-            amazonS3Client.deleteObject(new DeleteObjectRequest(AWS_BUCKET_NAME, post.getImageId()));
-        }
-
-        PutObjectResult putObjectResult = amazonS3Client.putObject(new PutObjectRequest
-                (AWS_BUCKET_NAME, "spring_microservice/"+fileName, convertedFile)
+        AwsSnsPostDto postDetails = new AwsSnsPostDto(createdPost.getId(),
+                request.getTitle(),
+                "Create",
+                "Hi there you have created new post");
+        Gson gson = new Gson();
+        PublishResult publishResult = amazonSNSClient.publish(new PublishRequest
+                (snsArn, gson.toJson(postDetails), "Post Activity")
         );
 
-        System.out.println("Put result => versionId : " + putObjectResult.getVersionId() + "\n"
-        + "contentMD5 : " + putObjectResult.getContentMd5() + "\n"
-        + "metadata : " + putObjectResult.getMetadata() + "\n"
-        + "ETag : " + putObjectResult.getETag());
+        System.out.println("Publish result => " + publishResult.getMessageId());
 
-        updatePost(postId, updatePostDto);
-    }
-
-    public DownloadFileDto downloadFile(String id) throws IOException {
-        Posts post = getPost(id);
-        String filePath = post.getImageId();
-        String fileName = post.getImageId().split("/")[1];
-
-        System.out.println("File path : " + post.getImageId() + " fileName : " + fileName);
-
-        S3Object s3Object = amazonS3Client.getObject(new GetObjectRequest(AWS_BUCKET_NAME, filePath));
-
-        System.out.println("S3Object => " + s3Object.getBucketName() + " key : " + s3Object.getKey());
-
-        return new DownloadFileDto(s3Object.getObjectContent().readAllBytes(), fileName, s3Object.getObjectMetadata().getContentType());
+        return createdPost.getId();
     }
 
     public Posts getPost(String id){
@@ -121,6 +99,19 @@ public class PostsService {
     }
 
     public void deletePost(String id){
-        postsRepository.deleteById(id);
+        Posts post = postsRepository.findById(id).orElse(null);
+
+        if (post != null){
+            postsRepository.deleteById(id);
+            AwsSnsPostDto postDetails = new AwsSnsPostDto(post.getId(),
+                    post.getTitle(),
+                    "Delete",
+                    "Hi there you have deleted post");
+            Gson gson = new Gson();
+            PublishResult publishResult = amazonSNSClient.publish(new PublishRequest
+                    (snsArn, gson.toJson(postDetails))
+            );
+            System.out.println("Publish result => " + publishResult.getMessageId());
+        }
     }
 }
